requirements.txt
blinker==1.6.2
click==8.1.3
Flask==2.3.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
Werkzeug==2.3.4
python-dotenv>=0.19.2
authlib>=1.0
requests>=2.27.1
bs4
